Disable Node type stripping when running tests
In a recent version, Node supports running TypeScript directly. It
does this by a simple compilation step where it just strips the types
and doesn't do anything more. It doesn't support ant TypeScript
features that requires new JavaScript code to be emitted. This is not
compatible with how we're running the tests. Therefore we disable
this feature when running the tests.